fix: resolve eBPF compilation error and type mismatches v1.3.1

+++ backend/services/flood.go
@@ -99,7 +99,7 @@ func (fp *FloodProtection) CheckIP(ip string, packetCount int, byteCount int64) 
 			if tracker.Violations >= thresholds.MaxViolations {
 				tracker.Blocked = true
 				tracker.BlockedUntil = time.Now().Add(thresholds.BlockDuration)
-				go fp.recordAttack(ip, "Connection Flood", int(tracker.PacketsPerSec))
+				go fp.recordAttack(ip, "Connection Flood", int64(tracker.PacketsPerSec))
 				return true
 			}
 		}
@@ -112,7 +112,7 @@ func (fp *FloodProtection) CheckIP(ip string, packetCount int, byteCount int64) 
 		if tracker.Violations >= thresholds.MaxViolations {
 			tracker.Blocked = true
 			tracker.BlockedUntil = time.Now().Add(thresholds.BlockDuration)
-			go fp.recordAttack(ip, "PPS Flood", tracker.PacketsPerSec)
+			go fp.recordAttack(ip, "PPS Flood", int64(tracker.PacketsPerSec))
 			return true
 		}
 	}
@@ -124,7 +124,7 @@ func (fp *FloodProtection) CheckIP(ip string, packetCount int, byteCount int64) 
 		if tracker.Violations >= thresholds.MaxViolations {
 			tracker.Blocked = true
 			tracker.BlockedUntil = time.Now().Add(thresholds.BlockDuration)
-			go fp.recordAttack(ip, "Bandwidth Flood", tracker.PacketsPerSec)
+			go fp.recordAttack(ip, "Bandwidth Flood", int64(tracker.PacketsPerSec))
 			return true
 		}
 	}
@@ -208,7 +208,7 @@ func (fp *FloodProtection) UnblockIP(ip string) {
 
 // recordAttack logs attack event to DB and sends webhook alert
 // This should be called as a goroutine to avoid blocking the main packet processing path
-func (fp *FloodProtection) recordAttack(ip string, attackType string, pps int) {
+func (fp *FloodProtection) recordAttack(ip string, attackType string, pps int64) {
 	// 1. Resolve Country
 	countryName := "Unknown"
 	countryCode := "XX"
